Fixed serious problem in which it segfaulted when the CSV file uses ;
The lovely Excel uses semicolons instead of commas in its generated CSV
files. As my only test was using the CSVs generated by OpenOffice, no
problem was observed.

When Deborah tried to test it, the program crashed, because of a quite
stupid check performed by the code:

  refclass = strrchr(line, ',') + 1;
  if (!refclass) {
       ...

So the code didn't properly notify the wrong input format nor handled it
graciously.